+++ google-beta/resource_filestore_instance_generated_test.go
@@ -20,9 +20,10 @@ import (
 
 	"github.com/hashicorp/terraform/helper/acctest"
 	"github.com/hashicorp/terraform/helper/resource"
+	"github.com/hashicorp/terraform/terraform"
 )
 
-func TestAccFilestoreInstance_filestoreInstanceBasicExample(t *testing.T) {
+func TestAccFilestoreInstance_FilestoreInstanceBasicExample(t *testing.T) {
 	t.Parallel()
 
 	resource.Test(t, resource.TestCase{
@@ -31,7 +32,7 @@ func TestAccFilestoreInstance_filestoreInstanceBasicExample(t *testing.T) {
 		CheckDestroy: testAccCheckFilestoreInstanceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccFilestoreInstance_filestoreInstanceBasicExample(acctest.RandString(10)),
+				Config: testAccFilestoreInstance_FilestoreInstanceBasicExample(acctest.RandString(10)),
 			},
 			{
 				ResourceName:            "google_filestore_instance.instance",
@@ -43,7 +44,7 @@ func TestAccFilestoreInstance_filestoreInstanceBasicExample(t *testing.T) {
 	})
 }
 
-func testAccFilestoreInstance_filestoreInstanceBasicExample(val string) string {
+func testAccFilestoreInstance_FilestoreInstanceBasicExample(val string) string {
 	return fmt.Sprintf(`
 resource "google_filestore_instance" "instance" {
   name = "test-instance-%s"
@@ -62,4 +63,26 @@ resource "google_filestore_instance" "instance" {
 }
 `, val,
 	)
+}
+
+func testAccCheckFilestoreInstanceDestroy(s *terraform.State) error {
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "google_filestore_instance" {
+			continue
+		}
+
+		config := testAccProvider.Meta().(*Config)
+
+		url, err := replaceVarsForTest(rs, "https://file.googleapis.com/v1beta1/projects/{{project}}/locations/{{zone}}/instances/{{name}}")
+		if err != nil {
+			return err
+		}
+
+		_, err = sendRequest(config, "GET", url, nil)
+		if err == nil {
+			return fmt.Errorf("FilestoreInstance still exists at %s", url)
+		}
+	}
+
+	return nil
 }
