+++ google-beta/resource_filestore_instance_test.go
@@ -2,12 +2,10 @@ package google
 
 import (
 	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform/helper/acctest"
 	"github.com/hashicorp/terraform/helper/resource"
-	"github.com/hashicorp/terraform/terraform"
 )
 
 func TestAccFilestoreInstance_basic(t *testing.T) {
@@ -60,31 +58,6 @@ func TestAccFilestoreInstance_update(t *testing.T) {
 			},
 		},
 	})
-}
-
-func testAccCheckFilestoreInstanceDestroy(s *terraform.State) error {
-	config := testAccProvider.Meta().(*Config)
-
-	for _, rs := range s.RootModule().Resources {
-		if rs.Type != "google_filestore_instance" {
-			continue
-		}
-
-		redisIdParts := strings.Split(rs.Primary.ID, "/")
-		if len(redisIdParts) != 3 {
-			return fmt.Errorf("Unexpected resource ID %s, expected {project}/{region}/{name}", rs.Primary.ID)
-		}
-
-		project, region, inst := redisIdParts[0], redisIdParts[1], redisIdParts[2]
-
-		name := fmt.Sprintf("projects/%s/locations/%s/instances/%s", project, region, inst)
-		_, err := config.clientFilestore.Projects.Locations.Get(name).Do()
-		if err == nil {
-			return fmt.Errorf("Filestore instance still exists")
-		}
-	}
-
-	return nil
 }
 
 func testAccFilestoreInstance_basic(name string) string {
